+++ ElkaUWP.DataLayer/Usos/Entities/PostalAddresses.cs
@@ -1,0 +1,21 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ElkaUWP.DataLayer.Usos.Entities
+{
+    public class PostalAddresses
+    {
+        [JsonProperty("type")]
+        public string Type { get; set; }
+
+        [JsonProperty("address")]
+        public string Address { get; set; }
+
+        [JsonProperty("type_name")]
+        public LangDict TypeName { get; set; }
+    }
+}
