+++ ElkaUWP.DataLayer/Usos/Entities/PostalAddressesContainer.cs
@@ -14,15 +14,5 @@ namespace ElkaUWP.DataLayer.Usos.Entities
 
     }
 
-    public class PostalAddresses
-    {
-        [JsonProperty("type")]
-        public string Type { get; set; }
-
-        [JsonProperty("address")]
-        public string Address { get; set; }
-
-        [JsonProperty("type_name")]
-        public LangDict TypeName { get; set; }
-    }
+   
 }
